Fix budget endpoints to match actual Runtime API paths
- budget/quota.go: Use /api/v2/organizations/{org}/members/{user}/agent-quota
  (actual Runtime path) instead of placeholder /api/v2/agents/quota/{ownerID}
- budget/cost.go: Use /api/v2/deployment/cost (now exists in Runtime)
  and align CostSummary field types to int64
- identity/agents.go: Fix GetAgentQuota to use correct org+user path

+++ budget/cost.go
@@ -11,9 +11,9 @@ import (
 
 // CostSummary contains aggregated cost information.
 type CostSummary struct {
-	TotalCredits     int `json:"total_credits"`
-	UsedCredits      int `json:"used_credits"`
-	RemainingCredits int `json:"remaining_credits"`
+	TotalCredits     int64 `json:"total_credits"`
+	UsedCredits      int64 `json:"used_credits"`
+	RemainingCredits int64 `json:"remaining_credits"`
 }
 
 // GetCostSummary returns the cost summary for the deployment.
